Roll back a Contact Flow
Roll back a Contact Flow not the same in all Regions as the London does not have the latest version of the Connect Web Application (Portal) and so there is no Drop-down list to select different versions, only the ability to revert to the Last Saved or Last Published. Also the View historical changes link at the bottom of the Contact Flows page just shows "Coming soon".